WIP
Seems to have some issues when the trays load is to high. There might be some kind of death lock in the events structure that does no allow the system continue when its full of trays. Appears a bottle neck where all the trays are in the input of a stopper but no one is able to move because the next stopper is full. Maybe it is not a problem for the normal behavior of the system. 

It is important to consider that the problem may be intrinsic to the discrete events system as it is designed, and not a cause of a bug in the code. Any way, it may be interesting to take a look. There is more information on wandb.

There is also another issue with the wandb loggin. It sometimes takes a lot of time to call wandb.log(). Its strange, because, theoretically, that function only stores the data locally until the end of the run. Maybe is an issue with the access for the file system (it stores data on files as I found on the docs), or meybe is an issue with the data adquisition (some kind of weird recursive access to the simulation data? there are A LOT of circular references and it has to be managed with care)