WIP
Halting problem detected. Taken one linear path on the system, the path has a maximum number of trays of N-1, being N the number of stopper in the path, because it needs a free stopper to start a move. The inputs at DIR04 must be ignored if there is a tray already on the system, to avoid that issue. It there is a group of trays that generate a ciclic graph with N-1 trays, so it is full, if in any bifurcation there is a tray outside that path that wants to enter the on that path, and has more priority (its upper on the configuration) it would enter the path instead of letting the path flow, causing a halt of the state machines. In the trays entrance is easier to fix, because we just want to reduce the priority of the non important path. But in other places that can't be done, because several different ciclic paths can be generated, and the priorities to avoid the halt at each of them are different. 
So, a halting detector could be implemented, to allow bifurcations to avoid that issue.
But idk if that is really important, because these cicles are likely to be generated on the test context, were all path take the same time (to simplify the configuration) and that created the synchronization required to achieve that N-1 occupancy. Also, maybe, just more smart priorities, which could check the queue length at its sourroundings could be enought to solve that problem.