Fix typing issues and update cycles algorithm
Now the detector generates a dict for each stopper that is part of a cycle and has to lock tangent stoppers. It contains  an array for each output with the cycles that have to be analyzed. Cycles will be converted to classes that will implement the logic to analyze the cycle population to avoid issues.